Fix battinfo for 2-digit hours remaining
Gnarly. 10 hours of estimated battery life.  Prior to this, the batt.rb
script used for the tmux status line only expected a single digit for
hours of battery life remaining, causing the regex to not match, thus
not reporting life stats.  This commit updates the regex to support
multiple digits in the hour.